agrego libreria de mapa y agrego funcion estacercade otro poi

+++ src/ParadaColectivo.java
@@ -3,7 +3,7 @@ public class ParadaColectivo extends POI {
 	private String empresa;
 	private int numeroDeLinea;
 	
-	public ParadaColectivo(String nombre, float latitud, float longitud) {
+	public ParadaColectivo(String nombre, double latitud, double longitud) {
 		super.nombre = nombre;
 		super.latitud = latitud;
 		super.longitud = longitud;
